splitted module cropping into two steps, (i) estimation of module quadrilaterals, (ii) cropping of modules
