+++ TODO.txt
@@ -3,7 +3,6 @@ TODO:
 - remove preview images to make user experience easier (modify view_gps.py which uses preview images), update docs (using your own data)
 - remove split_tiffs and rotate step and move into scripts directory (also put split_seqs there), update docs accordingly (using your own data)
 - move cropping after mapping and consider merged modules directly in the cropping, so that we can prevent copying the patches folder to patches_final
-- remove segmentation/preview (or make an option to omit this output)
 - zip the patches of each module using Python ZipFile
