+++ defaults.yml
@@ -11,10 +11,6 @@ rotate_frames:
 segment_pv_modules:
   output_video_fps: 8.0  # [Hz]
 
-crop_and_rectify_modules:
-  min_iou: 0.9
-  rotate_mode: portrait
-
 track_pv_modules:
   motion_model: homography  # how to model the motion between two subsequent frames
   orb_nfeatures: 5000  # number of ORB features to extract in each frame (needed to compute visual distance)
@@ -24,6 +20,13 @@ track_pv_modules:
   match_distance_thres: 20.0  # maximum feature distance of two feature descriptors to be matched (needed to compute visual distance)
   max_distance: 60  # [px] maximum Euclidean distance a module center point can travel in two subsequent frames to still be considered the same module  
   output_video_fps: 8.0  # [Hz]
+
+#crop_and_rectify_modules:
+#  min_iou: 0.9
+#  rotate_mode: portrait
+
+compute_pv_module_quadrilaterals:
+  min_iou: 0.9
 
 prepare_opensfm:
   select_frames_mode: gps  # select frames for 3D reconstruction based on GPS alone ("gps") or GPS and visual cues ("gps_visual")
@@ -59,4 +62,7 @@ refine_triangulation:
   max_num_modules: 300  # if number of modules per frame exceeds this value skip do not consider this frame for refinement
   optimizer_steps: 10  # number of graph optimization steps to perform
 
-reorganize_patches:
+#reorganize_patches:
+
+crop_pv_modules:
+  rotate_mode: portrait
